fix: rename recent-posts to recent-card
