+++ components/section/postListSection/index.tsx
@@ -2,8 +2,8 @@ import { NextPage } from 'next'
 import HighlightSection from '../../highlight-card'
 import LatestCard from '../../latest-card'
 import ListPost from '../../list-post'
-import RecentPostCard from '../../recent-posts'
-import RecentPostsContainer from '../../recent-posts/container'
+import RecentPostCard from '../../recent-card'
+import RecentPostsContainer from '../../recent-card/container'
 
 import { PostSectionFixtures } from './constants'
 import { PostSectionProps } from './types'
